refactor: NotFoundPointOwnerException, PointErrors 패키징 이동
